refactor: 피드 의존성 정리
