+++ backend/src/main/java/com/yigongil/backend/domain/feedpost/FeedService.java
@@ -2,8 +2,8 @@ package com.yigongil.backend.domain.feedpost;
 
 import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.study.Study;
+import com.yigongil.backend.domain.study.StudyRepository;
 import com.yigongil.backend.request.FeedPostCreateRequest;
-import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,24 +11,24 @@ import org.springframework.transaction.annotation.Transactional;
 public class FeedService {
 
     private final FeedPostRepository feedPostRepository;
+    private final StudyRepository studyRepository;
 
-    public FeedService(FeedPostRepository feedPostRepository) {
+    public FeedService(FeedPostRepository feedPostRepository, StudyRepository studyRepository) {
         this.feedPostRepository = feedPostRepository;
+        this.studyRepository = studyRepository;
     }
 
     @Transactional
-    public void createFeedPost(Member member, Study study, FeedPostCreateRequest request) {
+    public void createFeedPost(Member member, Long studyId, FeedPostCreateRequest request) {
+        Study study = studyRepository.getById(studyId);
+
         FeedPost regularFeedPost = FeedPost.builder()
                                            .author(member)
                                            .study(study)
                                            .imageUrl(request.imageUrl())
                                            .content(request.content())
                                            .build();
-        feedPostRepository.save(regularFeedPost);
-    }
 
-    @Transactional(readOnly = true)
-    public List<FeedPost> findFeedPosts(Long studyId, Long oldestFeedPostId) {
-        return feedPostRepository.findAllByStudyIdStartWithOldestFeedPostId(studyId, oldestFeedPostId);
+        feedPostRepository.save(regularFeedPost);
     }
 }
